requirements.txt
flask
gunicorn
requests
psycopg2
psycopg2-binary
pyyaml
